test(stores): split StoresWrapperTests to satisfy 500-line cap (F1, #261 remediation cycle 1)
Remediation cycle 1 for the feature-review R1 blocking finding: StoresWrapperTests.cs
exceeded the 500-line file-size limit (688 lines) after F1 added disabled-store tests.

- Extract the F1 disabled-store tests plus the InclusionFilters_* group and the
  AssertInclusionDecision helper into a new StoresWrapperDisableTests.cs (368 lines),
  wired into UtilitiesCS.Test.csproj. StoresWrapperTests.cs now 415 lines. Both < 500.
  Test methods moved verbatim; no assertion or logic change.
- N1: convert the two ReenableAsync guard tests to async Task and await their
  ThrowAsync assertions so the guard paths are actually verified.

Toolchain green; repo line coverage 81.61% (>= 80%); no regression. Includes the
initial and reaudit-input audit artifacts and cycle-1 QA evidence.
