+++ docs/features/active/2026-07-07-store-disable-service-261/evidence/qa-gates/qa-02-analyzers-cycle1.md
@@ -1,0 +1,27 @@
+# QA Gate 2 — .NET Analyzers Build (Remediation Cycle 1)
+
+- Timestamp: 2026-07-08T00-45
+- Command: `msbuild TaskMaster.sln /t:Build /p:Configuration=Debug /p:Platform="Any CPU" /p:EnableNETAnalyzers=true /p:EnforceCodeStyleInBuild=true`
+  (invoked via the MSBuild.exe full path found under Visual Studio 18 Community, using
+  single-dash-converted-to-slash-equivalent switches; see Deviation note)
+- EXIT_CODE: 0
+- Output Summary: **Build succeeded.** 20 Warning(s), 0 Error(s). All 20 warnings are
+  pre-existing, in files unrelated to this remediation (CS8632 nullable-annotation-context
+  warnings in `OlTableExtensions_Tests.cs`, `ProgressTracker_Tests.cs`,
+  `ConversationHelper_ExtendedTests.cs`, `ManualFireTimerWrapper.cs`; CS0067 unused-event
+  warnings in `SmartSerializable_Tests.cs`, `SmartSerializableBase_Tests.cs`,
+  `StoreWrapperControllerTests.cs`). Zero analyzer warnings or errors on the three touched files
+  (`StoresWrapperTests.cs`, `StoresWrapperDisableTests.cs`, `StoreDisableServiceTests.cs`),
+  confirmed via `grep -i "StoresWrapper\|StoreDisableService"` against the full build log
+  (no matches).
+
+## Deviation Note
+
+- MSBuild invocation form: the bare `msbuild` command is not on this git-bash session's `PATH`
+  (`msbuild: command not found`), even though `where msbuild` (Windows-native lookup) resolves it
+  to `C:\Program Files\Microsoft Visual Studio\18\Community\MSBuild\Current\Bin\MSBuild.exe`. The
+  full path to `MSBuild.exe` was invoked directly with equivalent single-dash switches
+  (`-t:Build -p:Configuration=Debug -p:Platform="Any CPU" -p:EnableNETAnalyzers=true
+  -p:EnforceCodeStyleInBuild=true`), which MSBuild treats identically to the `/`-prefixed forms.
+  This is an environment PATH-resolution workaround, not a change to the build target, properties,
+  or semantics specified by the plan/policy.
