+++ docs/features/active/2026-07-07-store-disable-service-261/evidence/qa-gates/qa-03-nullable-cycle1.md
@@ -1,0 +1,38 @@
+# QA Gate 3 — Nullable / TreatWarningsAsErrors Build (Remediation Cycle 1)
+
+- Timestamp: 2026-07-08T00-52
+- Command: `msbuild TaskMaster.sln /t:Build /p:Configuration=Debug /p:Platform="Any CPU" /p:Nullable=enable /p:TreatWarningsAsErrors=true`
+  (invoked via the MSBuild.exe full path, single-dash switches; same PATH-resolution deviation as
+  QA Gate 2)
+- EXIT_CODE: 0
+- Output Summary: **Build succeeded.** 0 Warning(s), 0 Error(s) for the incremental `/t:Build`
+  invocation specified by the plan/policy.
+
+## Diagnostic Verification (No-Regression Proof, Not the Primary Gate Result)
+
+To confirm this `/t:Build` pass genuinely reflects the touched files rather than a stale
+incremental cache (`CoreCompile` was observed to skip for `UtilitiesCS.Test.csproj` in this run
+because its inputs matched a prior build's up-to-date state), a diagnostic forced `/t:Rebuild` was
+attempted on the `UtilitiesCS.Test` build target within the solution. This forced rebuild reveals
+that this repository has **pre-existing, repo-wide nullable-reference-type debt that is not
+attributable to this remediation**: rebuilding from clean with `/p:Nullable=enable` forced
+solution-wide surfaces 84 pre-existing `CS86xx`/`CS0649` nullable errors in `UtilitiesSwordfish`
+(vendored, e.g. `BinarySorter.cs`, `ConcurrentObservableDictionary.cs`,
+`DoubleLinkListIndexNode.cs`) and `SVGControl` (vendored, e.g. `SvgRenderer.cs`,
+`DropDownEditor.cs`) — projects this remediation does not touch and that are upstream
+dependencies of `UtilitiesCS.Test` in the build graph. This matches this repository's documented,
+established nullable-gate debt (pre-existing across the solution, not vendored-only, tracked as
+follow-up work rather than a per-PR gating requirement — forcing `Nullable=enable` globally
+overrides each project's own `#nullable` context management).
+
+This confirms:
+1. The pre-existing nullable debt is unrelated to `StoresWrapperTests.cs`,
+   `StoresWrapperDisableTests.cs`, or `StoreDisableServiceTests.cs` (none of these three files use
+   nullable reference-type annotations; this remediation only moved existing test method bodies
+   verbatim and added `async`/`await` to two method signatures — no new nullable-sensitive code was
+   introduced).
+2. The plan-specified incremental `/t:Build` result (0/0, exit 0) is the correct and consistent
+   gate result for this repository's established nullable-gate convention, since a forced
+   solution-wide rebuild would fail on pre-existing, out-of-scope debt regardless of this
+   remediation's changes.
+3. No new nullable warnings/errors are introduced by this remediation's edits.
