+++ docs/features/active/2026-07-07-store-disable-service-261/evidence/qa-gates/qa-05-coverage-post-change-cycle1.md
@@ -1,0 +1,27 @@
+# QA Gate 5 — Repo-Wide Post-Change Coverage (Remediation Cycle 1)
+
+- Timestamp: 2026-07-08T00-52
+- Command: `pwsh ./scripts/vscode/Invoke-MSTestWithCoverage.ps1 -CoverageOutput coverage/remediation-cycle1-post-change.cobertura.xml`
+  (forward-slash path form; see the identical deviation rationale in
+  `evidence/remediation-baseline/test-coverage-baseline-cycle1.md`)
+- EXIT_CODE: 1 (non-zero; the script's own gate treats any test failure as a hard failure and
+  throws — see Findings below; the single failure is the same pre-existing, environment-dependent
+  test identified in the P0-T8 baseline)
+- Output Summary:
+  - Total tests: 5032 (unchanged from P0-T8 baseline)
+  - Passed: 5031 (unchanged from P0-T8 baseline)
+  - Failed: 1 — `LiveHookup_OnSta_CompletesAndDoesNotBlockStaBeyondThreshold`, confirmed the same
+    pre-existing live-Outlook-COM-dependent failure as the P0-T8 baseline (unrelated to this
+    remediation's touched files)
+  - Total test time: 48.4400 seconds
+  - Repo-wide line coverage (Cobertura top-level `line-rate`, same method as the P0-T8 baseline):
+    **81.61%** (`lines-covered="119396"` / `lines-valid="146294"` = 0.8161373672194349), versus
+    the baseline's 81.62% (`119363` / `146244` = 0.8161907497059708). The denominator increased by
+    50 lines and the numerator by 33 lines — consistent with the R1 split producing a new file
+    (`StoresWrapperDisableTests.cs`) whose duplicated helper methods add a small number of
+    additional executable lines beyond the moved-verbatim test bodies, all of which execute when
+    the tests run. The 0.01-percentage-point difference is not a regression and remains
+    comfortably above the CLAUDE.md 80% floor.
+  - New file coverage: the `UtilitiesCS.Test.OutlookObjects.Store.StoresWrapperDisableTests` class
+    reports `line-rate="1"` (100%) in the post-change Cobertura report — every line in the new
+    test file executes when its tests run, confirming no coverage loss from the split.
