+++ docs/features/active/2026-07-07-store-disable-service-261/evidence/qa-gates/qa-07-file-size-final-cycle1.md
@@ -1,0 +1,16 @@
+# QA Gate 7 — Final File Size Verification (Remediation Cycle 1)
+
+- Timestamp: 2026-07-08T00-55
+- Commands (corrected `.Count` form; see
+  `evidence/remediation-baseline/wc-stores-wrapper-tests-before.md` for the
+  `Measure-Object -Line` blank-line undercount rationale):
+  - `(Get-Content 'UtilitiesCS.Test\OutlookObjects\Store\StoresWrapperTests.cs').Count`
+  - `(Get-Content 'UtilitiesCS.Test\OutlookObjects\Store\StoresWrapperDisableTests.cs').Count`
+- EXIT_CODE: 0 (both commands)
+- Output Summary:
+  - `StoresWrapperTests.cs`: **415** lines (unchanged from the post-split P1-T7 measurement;
+    CSharpier's format pass, run in QA Gate 1, made no changes to this file).
+  - `StoresWrapperDisableTests.cs`: **368** lines (unchanged from the post-split P1-T8
+    measurement; CSharpier's format pass made no changes to this file).
+  - Both files are comfortably under the 500-line cap (CLAUDE.md §4.1 /
+    `.claude/rules/general-code-change.md` file-size limit).
